Fixup boot code and add comments to boot code and kernel.zig
What was happening was that the `kmain` entry-point was never actually
getting invoked due to an issue with the boot code that a newer version
of QEMU revealed. I believe what was happening was the timer interrupts
were being enabled in the boot code and before reaching kmain, we were
actually getting an interrupt and since the trap code does not currently
handle interrupts, the interrupts were just continuously being delivered.